+++ modules/alloy/src/main/ts/ephox/alloy/ui/schema/DropdownSchema.ts
@@ -35,7 +35,13 @@ const schema = Fun.constant([
 const parts: () => PartType.PartTypeAdt[] = Fun.constant([
   PartType.external<DropdownDetail, DropdownSpec>({
     schema: [
-      Fields.tieredMenuMarkers()
+      Fields.tieredMenuMarkers(),
+      // Defining a defaulted field isn't necessary when dealing with
+      // external parts, because the post-boulder part spec is not passed
+      // through to any of these functions (defaults, overrides etc.). So all
+      // this does is make it a bit clearer what you should expect, but remember
+      // that the default value here is irrelevant!
+      FieldSchema.defaulted('fakeFocus', false)
     ],
     name: 'menu',
     defaults: (detail) => {