+++ modules/tinymce/src/themes/silver/main/ts/ui/core/complex/BespokeSelect.ts
@@ -138,7 +138,15 @@ const generateSelectItems = (_editor: Editor, backstage: UiFactoryBackstage, spe
   const getFetch = (backstage: UiFactoryBackstage, getStyleItems: () => FormatItem[]) => (comp: AlloyComponent, callback: (menu: Optional<TieredData>) => void) => {
     const preItems = getStyleItems();
     const items = validateItems(preItems);
-    const menu = NestedMenus.build(items, ItemResponse.CLOSE_ON_EXECUTE, backstage, false);
+    const menu = NestedMenus.build(
+      items,
+      ItemResponse.CLOSE_ON_EXECUTE,
+      backstage,
+      {
+        isHorizontalMenu: false,
+        search: Optional.none()
+      }
+    );
     callback(menu);
   };