+++ modules/tinymce/src/themes/silver/main/ts/ui/general/Button.ts
@@ -179,6 +179,8 @@ export const renderFooterButton = (spec: FooterButtonSpec, buttonType: string, b
     const fixedSpec: Toolbar.ToolbarMenuButton = {
       ...spec,
       type: 'menubutton',
+      // Currently, dialog-based menu buttons cannot be searchable.
+      search: Optional.none(),
       onSetup: (api) => {
         api.setEnabled(spec.enabled);
         return Fun.noop;