+++ modules/tinymce/src/themes/silver/main/ts/ui/menus/menu/MenuConversion.ts
@@ -82,17 +82,37 @@ const generateValueIfRequired = (item: Menu.NestedMenuItemSpec): Menu.NestedMenu
 
 // Takes items, and consolidates them into its return value
 const expand = (items: string | Array<string | SingleMenuItemSpec>, menuItems: MenuItemRegistry): ExpandedMenus => {
+  // Fistly, we do all substitution using the registry for any items referenced by their
+  // string key.
   const realItems = unwrapReferences(Type.isString(items) ? items.split(' ') : items, menuItems);
+
+  // Now that we have complete bridge Item specs for all items, we need to collect the
+  // submenus, items in the primary menu, and triggering menu items all into one
+  // giant object to from the building blocks on our TieredData
   return Arr.foldr(realItems, (acc, item) => {
     if (isExpandingMenuItem(item)) {
+      // We generate a random value for item, but only if there isn't an existing value
       const itemWithValue = generateValueIfRequired(item);
+
+      // The newData isn't quite in the format you might expect. The list of items
+      // for an item with nested items is just the single parent item. All of the nested
+      // items becomes part of '.menus'. Finally, the expansions is just a map from
+      // the triggering item to the first submenu. Incidentally, they are given the same
+      // value (triggering item and submenu), for convenience.
       const newData = getFromExpandingItem(itemWithValue, menuItems);
       return {
+        // Combine all of our current submenus and items with the new submenus created by
+        // this item with nested subitems
         menus: Merger.deepMerge(acc.menus, newData.menus),
+        // Add our parent item into the list of items in the *current menu*.
         items: [ newData.item, ...acc.items ],
+        // Merge together our "this item opens this submenu" objects
         expansions: Merger.deepMerge(acc.expansions, newData.expansions)
       };
     } else {
+      // If we aren't creating any submenus, then all we need to do is add this item
+      // to the list of items in the current menu. So this is the same as an expanding
+      // menu item, except it doesn't add to `menus` or `expansions`.
       return {
         ...acc,
         items: [ item, ...acc.items ]