+++ modules/boulder/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Added an optional parameter to `oneOf` that allows boulder to generate consistent output for different schema. #TINY-8952
+
 ## 7.0.0 - 2022-03-03
 
 ### Changed